pyproject.toml

[build-system]
requires = ["setuptools"] 
build-backend = "setuptools.build_meta"

[project]
name = "deafrica-water-quality"
version = "0.0.1"
authors = [{name = "Digital Earth Africa", email = "systems@digitalearthafrica.org"}]
maintainers = [{name = "Digital Earth Africa", email = "systems@digitalearthafrica.org"}]
description = "Helpful scripts to make water quality data available in a ODC deploymnent"
readme =  {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 1 - Planning",
    'Intended Audience :: Science/Research',
    'Topic :: Scientific/Engineering :: GIS',
    'License :: OSI Approved :: Apache Software License',
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "click",
    "datacube[performance,distributed,s3]",
    "deafrica-tools",
    "fsspec",
    "gcsfs",
    "odc-cloud[ASYNC]",
    "odc-dscache",
    "odc-stats",
    "requests",
    "s3fs",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "yamlint", 
 ]
test = [
    "pytest",
    "coverage"
]

[project.urls]
Repository = "https://github.com/digitalearthafrica/deafrica_water_quality.git"
"Bug Tracker" = "https://github.com/digitalearthafrica/deafrica_water_quality/issues"

[project.scripts]
wq-list-test-areas = "water_quality.cli.list_test_areas:cli"
wq-generate-tiles = "water_quality.cli.generate_tiles:cli"
wq-generate-tasks = "water_quality.cli.generate_tasks:cli"
wq-process-tasks = "water_quality.cli.process_wq_variables:cli"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
water_quality = ["data/*.parquet"]

[tool.ruff]
line-length = 79

[tool.pytest.ini_options]
addopts = "-s"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
